+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.16 - In progress
 ### Added
 - Different search configs for different types.
+- Link to derived Samples and Datasets.
 ### Changed
 - Fix VisTabs panel overflow.
 - Fix NoticeAlert to only display when errors exist.