+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.16 - In progress
 ### Added
 - Sort by last_modified.
+- Hide entity_type facet.
 - Different search configs for different types.
 ### Changed
 - Use the code that had been in portal-search directly.