+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.16 - In progress
 ### Added
+- Sort by last_modified.
 - Hide entity_type facet.
 - Different search configs for different types.
 ### Changed